pyproject.toml
[project]
name = "invest-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.6",
    "langchain-openai>=0.3.27",
    # RAG 관련 패키지들
    "langchain-community>=0.2.0",
    "langchain-text-splitters>=0.2.0",
    "langchain-experimental>=0.2.0",
    "langchain-chroma>=0.2.0",
    "langchain-huggingface>=0.2.0",
    "langchain-ollama>=0.2.0",
    "langchain-qdrant>=0.2.0",
    "langchain-pinecone>=0.2.0",
    # 문서 처리 패키지들
    "pypdf>=4.0.0",
    "beautifulsoup4>=4.12.0",
    "tiktoken>=0.6.0",
    # 임베딩 및 벡터 저장소 패키지들
    "transformers>=4.36.0",
    "sentence-transformers>=2.2.0",
    "chromadb>=0.4.0",
    "qdrant-client>=1.7.0",
    # 기타 유틸리티 패키지들
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "gradio>=4.0.0",
    "posthog>=2.4.0,<6.0.0",
    "langfuse>=3.1.3",
    "ragas>=0.2.15",
    "rapidfuzz>=3.13.0",
    "openpyxl>=3.1.5",
    "langgraph>=0.5.2",
    "langgraph-cli[inmem]>=0.3.4",
    "crawl4ai>=0.7.0",
    "python-dotenv>=1.1.1",
    "langchain-neo4j>=0.4.0",
    "pinecone>=7.3.0",
    "streamlit>=1.47.0",
    "jq>=1.10.0",
    "langchain-tavily>=0.2.9",
    "openai>=1.97.0",
]

